pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lpm-headphone"
version = "0.1.0"
description = "Configurable lumped-parameter headphone modeling (small-signal) from voltage to pressure."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "matplotlib>=3.8",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "ruff>=0.6",
  "pyyaml>=6.0",   # only if you want YAML configs
]

[project.scripts]
lpm = "lpm.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
